Move get_policies into fsdp util
